Purit/backend/requirements.txt
numpy
scipy
librosa
soundfile
torch
torchaudio
pymongo
tqdm
python-dotenv
fastapi
uvicorn
python-multipart
